Reduce a lot of unnecessary calls to JourneyPlanner when on admin
When editing some setting on the admin page, the new settings is persisted to Firestore.
When something is edited on Firestore, this triggers a new onSnapshot call.
The new snapshot call sets a new Settings object, although nothing is really changed.
This triggers a bunch of hooks that fetches new data.

This PR fixes this by ignoring the onSnapshot setter when on admin